requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
passlib
bcrypt==3.2.0
python-jose[cryptography]
pydantic[email]
requests
python-multipart
alembic
